+++ app/src/main/java/com/agileburo/anytype/ui/page/modals/actions/ProfileIconActionMenuFragment.kt
@@ -1,9 +1,10 @@
 package com.agileburo.anytype.ui.page.modals.actions
 
 import android.Manifest
-import android.app.Activity.RESULT_OK
+import android.app.Activity
 import android.content.Intent
 import android.content.pm.PackageManager
+import android.content.res.ColorStateList
 import android.os.Bundle
 import android.provider.MediaStore
 import android.view.View
@@ -14,31 +15,42 @@ import androidx.core.os.bundleOf
 import androidx.lifecycle.Observer
 import androidx.lifecycle.ViewModelProviders
 import com.agileburo.anytype.R
+import com.agileburo.anytype.core_ui.extensions.avatarColor
+import com.agileburo.anytype.core_utils.ext.firstDigitByHash
 import com.agileburo.anytype.core_utils.ext.parsePath
 import com.agileburo.anytype.core_utils.ext.toast
 import com.agileburo.anytype.core_utils.ui.BaseFragment
 import com.agileburo.anytype.di.common.componentManager
-import com.agileburo.anytype.emojifier.Emojifier
+import com.agileburo.anytype.domain.common.Id
+import com.agileburo.anytype.domain.common.Url
 import com.agileburo.anytype.library_page_icon_picker_widget.ui.ActionMenuAdapter
-import com.agileburo.anytype.library_page_icon_picker_widget.ui.ActionMenuAdapter.Companion.OPTION_CHOOSE_EMOJI
-import com.agileburo.anytype.library_page_icon_picker_widget.ui.ActionMenuAdapter.Companion.OPTION_CHOOSE_RANDOM_EMOJI
-import com.agileburo.anytype.library_page_icon_picker_widget.ui.ActionMenuAdapter.Companion.OPTION_CHOOSE_UPLOAD_PHOTO
-import com.agileburo.anytype.library_page_icon_picker_widget.ui.ActionMenuAdapter.Companion.OPTION_REMOVE
 import com.agileburo.anytype.library_page_icon_picker_widget.ui.ActionMenuDivider
 import com.agileburo.anytype.presentation.page.picker.DocumentIconActionMenuViewModel
-import com.agileburo.anytype.presentation.page.picker.DocumentIconActionMenuViewModel.Contract
 import com.agileburo.anytype.presentation.page.picker.DocumentIconActionMenuViewModel.ViewState
 import com.agileburo.anytype.presentation.page.picker.DocumentIconActionMenuViewModelFactory
-import com.agileburo.anytype.ui.page.modals.DocumentEmojiIconPickerFragment
 import com.bumptech.glide.Glide
-import com.bumptech.glide.load.engine.DiskCacheStrategy
-import kotlinx.android.synthetic.main.action_toolbar_page_icon.*
+import kotlinx.android.synthetic.main.action_toolbar_profile_icon.*
 import javax.inject.Inject
 
-class DocumentIconActionMenuFragment : BaseFragment(R.layout.action_toolbar_page_icon),
+class ProfileIconActionMenuFragment : BaseFragment(R.layout.action_toolbar_profile_icon),
     Observer<ViewState> {
 
-    private val target: String
+    /**
+     * avatar image url
+     */
+    private val image: Url?
+        get() = arguments?.getString(IMAGE_KEY)
+
+    /**
+     * user name
+     */
+    private val name: String?
+        get() = arguments?.getString(NAME_KEY)
+
+    /**
+     * target (profile) id
+     */
+    private val target: Id
         get() = requireArguments()
             .getString(ARG_TARGET_ID_KEY)
             ?: throw IllegalStateException(MISSING_TARGET_ERROR)
@@ -59,6 +71,18 @@ class DocumentIconActionMenuFragment : BaseFragment(R.layout.action_toolbar_page
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
+        initialize()
+    }
+
+    override fun onChanged(state: ViewState) {
+        when (state) {
+            is ViewState.Exit -> exit()
+            is ViewState.Error -> toast(state.message)
+            is ViewState.Loading -> toast(getString(R.string.loading))
+        }
+    }
+
+    private fun initialize() {
         container.setOnClickListener { exit() }
         setIcon()
         setupLogoTranslation()
@@ -67,20 +91,18 @@ class DocumentIconActionMenuFragment : BaseFragment(R.layout.action_toolbar_page
     }
 
     private fun setIcon() {
-        arguments?.getString(EMOJI_KEY)?.let { unicode ->
-            Glide
-                .with(emojiIconImage)
-                .load(Emojifier.uri(unicode))
-                .diskCacheStrategy(DiskCacheStrategy.ALL)
-                .into(emojiIconImage)
-        }
-        arguments?.getString(IMAGE_KEY)?.let { url ->
+        image?.let { url ->
             Glide
                 .with(icon)
                 .load(url)
                 .centerInside()
                 .circleCrop()
                 .into(imageIcon)
+        } ?: apply {
+            imageText.text = name?.firstOrNull().toString()
+            val pos = name?.firstDigitByHash() ?: 0
+            icon.backgroundTintList = ColorStateList.valueOf(requireContext().avatarColor(pos))
+            imageIcon.setImageDrawable(null)
         }
     }
 
@@ -122,48 +144,23 @@ class DocumentIconActionMenuFragment : BaseFragment(R.layout.action_toolbar_page
             overScrollMode = OVER_SCROLL_NEVER
             adapter = ActionMenuAdapter(
                 options = intArrayOf(
-                    OPTION_CHOOSE_EMOJI,
-                    OPTION_CHOOSE_RANDOM_EMOJI,
-                    OPTION_CHOOSE_UPLOAD_PHOTO,
-                    OPTION_REMOVE
+                    ActionMenuAdapter.OPTION_CHOOSE_UPLOAD_PHOTO,
+                    ActionMenuAdapter.OPTION_REMOVE
                 )
             ) { option ->
                 when (option) {
-                    OPTION_CHOOSE_EMOJI -> {
-                        parentFragment?.childFragmentManager?.let { manager ->
-                            manager.popBackStack()
-                            DocumentEmojiIconPickerFragment.new(
-                                context = target,
-                                target = target
-                            ).show(manager, null)
-                        }
-                    }
-                    OPTION_REMOVE -> vm.onEvent(
-                        Contract.Event.OnRemoveEmojiSelected(
+                    ActionMenuAdapter.OPTION_REMOVE -> vm.onEvent(
+                        DocumentIconActionMenuViewModel.Contract.Event.OnRemoveEmojiSelected(
                             context = target,
                             target = target
                         )
                     )
-                    OPTION_CHOOSE_RANDOM_EMOJI -> vm.onEvent(
-                        Contract.Event.OnSetRandomEmojiClicked(
-                            context = target,
-                            target = target
-                        )
-                    )
-                    OPTION_CHOOSE_UPLOAD_PHOTO -> {
+                    ActionMenuAdapter.OPTION_CHOOSE_UPLOAD_PHOTO -> {
                         proceedWithImagePick()
                     }
                     else -> toast("Not implemented")
                 }
             }
-        }
-    }
-
-    override fun onChanged(state: ViewState) {
-        when (state) {
-            is ViewState.Exit -> exit()
-            is ViewState.Error -> toast(state.message)
-            is ViewState.Loading -> toast(getString(R.string.loading))
         }
     }
 
@@ -181,6 +178,15 @@ class DocumentIconActionMenuFragment : BaseFragment(R.layout.action_toolbar_page
         )
     }
 
+    private fun openGallery() {
+        Intent(
+            Intent.ACTION_PICK,
+            MediaStore.Images.Media.INTERNAL_CONTENT_URI
+        ).let { intent ->
+            startActivityForResult(intent, SELECT_IMAGE_CODE)
+        }
+    }
+
     private fun hasExternalStoragePermission() = ContextCompat.checkSelfPermission(
         requireActivity(),
         Manifest.permission.WRITE_EXTERNAL_STORAGE
@@ -188,29 +194,16 @@ class DocumentIconActionMenuFragment : BaseFragment(R.layout.action_toolbar_page
 
     override fun onActivityResult(requestCode: Int, resultCode: Int, data: Intent?) {
         super.onActivityResult(requestCode, resultCode, data)
-        if (resultCode == RESULT_OK && requestCode == SELECT_IMAGE_CODE) {
+        if (resultCode == Activity.RESULT_OK && requestCode == SELECT_IMAGE_CODE) {
             data?.data?.let { uri ->
                 val path = uri.parsePath(requireContext())
                 vm.onEvent(
-                    Contract.Event.OnImagePickedFromGallery(
+                    DocumentIconActionMenuViewModel.Contract.Event.OnImagePickedFromGallery(
                         context = target,
                         path = path
                     )
                 )
             }
-        }
-    }
-
-    private fun exit() {
-        parentFragment?.childFragmentManager?.popBackStack()
-    }
-
-    private fun openGallery() {
-        Intent(
-            Intent.ACTION_PICK,
-            MediaStore.Images.Media.INTERNAL_CONTENT_URI
-        ).let { intent ->
-            startActivityForResult(intent, SELECT_IMAGE_CODE)
         }
     }
 
@@ -222,17 +215,21 @@ class DocumentIconActionMenuFragment : BaseFragment(R.layout.action_toolbar_page
         componentManager().documentIconActionMenuComponent.release()
     }
 
+    private fun exit() {
+        parentFragment?.childFragmentManager?.popBackStack()
+    }
+
     companion object {
         fun new(
             y: Float?,
-            emoji: String?,
             image: String?,
+            name: String?,
             target: String
-        ): DocumentIconActionMenuFragment = DocumentIconActionMenuFragment().apply {
+        ): ProfileIconActionMenuFragment = ProfileIconActionMenuFragment().apply {
             arguments = bundleOf(
                 Y_KEY to y,
-                EMOJI_KEY to emoji,
                 IMAGE_KEY to image,
+                NAME_KEY to name,
                 ARG_TARGET_ID_KEY to target
             )
         }
@@ -240,8 +237,8 @@ class DocumentIconActionMenuFragment : BaseFragment(R.layout.action_toolbar_page
         private const val SELECT_IMAGE_CODE = 1
         private const val REQUEST_PERMISSION_CODE = 2
         private const val Y_KEY = "y"
-        private const val EMOJI_KEY = "emoji"
         private const val IMAGE_KEY = "image_key"
+        private const val NAME_KEY = "name_key"
         private const val ANIM_START_DELAY = 200L
         private const val ANIM_DURATION = 200L
         private const val ARG_TARGET_ID_KEY = "arg.picker.target.id"