+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### New features 🚀
 
 * Search-document-engine screen integrated on home dashboard and pages (#555)
+* Picking/removing avatar image for profile document (#568)
 
 ### Design & UX 🔳